backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
email-validator==2.2.0
python-dotenv==1.1.1
requests==2.32.3
beautifulsoup4==4.12.3
python-multipart==0.0.20
pypdf==5.4.0
python-docx==1.1.2
